refactor: BoardService에 테스트 더블 적용
