+++ src/main/java/chess/Application.java
@@ -1,6 +1,10 @@
 package chess;
 
 import chess.controller.ChessGameController;
+import chess.dao.PieceDAO;
+import chess.dao.SqlPieceDAO;
+import chess.dao.SqlTurnDAO;
+import chess.dao.TurnDAO;
 import chess.db.ConnectionManager;
 import chess.service.BoardService;
 import chess.view.InputView;
@@ -8,8 +12,12 @@ import chess.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
+        ConnectionManager connectionManager = new ConnectionManager("chess");
+        PieceDAO pieceDAO = new SqlPieceDAO(connectionManager);
+        TurnDAO turnDAO = new SqlTurnDAO(connectionManager);
+
         ChessGameController chessGameController = new ChessGameController(new OutputView(), new InputView(),
-                new BoardService(new ConnectionManager("chess")));
+                new BoardService(pieceDAO, turnDAO));
         chessGameController.run();
     }
 }
